% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_uwes.R
\name{calcular_uwes}
\alias{calcular_uwes}
\title{Calcula escores do UWES}
\usage{
calcular_uwes(
  df,
  itens_vigor,
  itens_dedicacao,
  itens_absorcao,
  total_itens,
  norma = "angst2009"
)
}
\description{
Função interna utilizada pelas versões UWES-9 e UWES-17.
}
\keyword{internal}
