% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificar.R
\name{classificar_uwes}
\alias{classificar_uwes}
\title{Classifica escores do UWES segundo normas disponiveis}
\usage{
classificar_uwes(vetor, dimensao, norma = "angst2009")
}
\arguments{
\item{vetor}{Vetor numerico de escores.}

\item{dimensao}{Dimensao avaliada.}

\item{norma}{Norma utilizada.}
}
\value{
Vetor de classificacoes.
}
\description{
Classifica escores do UWES segundo normas disponiveis
}
\keyword{internal}
