% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validacao.R
\name{validar_uwes}
\alias{validar_uwes}
\title{Valida dados do UWES}
\usage{
validar_uwes(df, total_itens)
}
\arguments{
\item{df}{data.frame contendo respostas.}

\item{total_itens}{Numero esperado de itens.}
}
\value{
Nenhum retorno. A funcao interrompe em caso de erro.
}
\description{
Valida dados do UWES
}
\keyword{internal}
